[Update] Removed functions from views and updated imports
